+++ ui/src/lib/community-helpers.js
@@ -87,6 +87,16 @@ export const PERMITTED_COMMUNITY_LINK_ATTRS = [
   'allowPublic'
 ]
 
+export const getGroupIds = {
+  query: gql`
+    query {
+      listGroups {
+        id
+      }
+    }
+  `
+}
+
 export const getMembers = id => ({
   query: gql`
     ${PublicProfileFieldsFragment}